refactor(lotto): 기존 Lotto 클래스 폴더 구조 변경
